Added an explicit cast because of glBinding

+++ src/nanovg_gl.h
@@ -387,7 +387,7 @@ static void glnvg__checkError(GLNVGcontext* gl, const char* str)
 	if ((gl->flags & NVG_DEBUG) == 0) return;
 	err = glGetError();
 	if (err != GL_NO_ERROR) {
-		printf("Error %08x after %s\n", err, str);
+		printf("Error %08x after %s\n", (unsigned int) err, str);
 		return;
 	}
 }
